requirements.txt
# Backend Requirements for Railway
# Newsletter Generation Engine dependencies

anthropic>=0.8.0
requests>=2.31.0
flask>=3.0.0
python-dotenv>=1.0.0
gunicorn>=21.2.0

# Note: Node.js and npm are required for the Brave Search MCP Server
# The MCP server will be installed via NPX during runtime
